Conversational-AI-Sales-Voice-Bot-/requirements.txt
Flask
pandas
numpy
elevenlabs
openai
langchain
scipy
sounddevice
wavio
python-dotenv
